add backend for mongodb

+++ nestjs-typescript/src/mysql/users/users.service.ts
@@ -11,8 +11,8 @@ export class UsersService {
   constructor(private readonly prisma: PrismaService) {}
 
   async create(createUserDto: CreateUserDto) {
-    const user = await this.prisma.users.findUnique({ where: { id: createUserDto.Username } });
-    const email = await this.prisma.users.findUnique({ where: { id: createUserDto.Email } });
+    const user = await this.prisma.users.findFirst({ where: { Username: createUserDto.Username } });
+    const email = await this.prisma.users.findFirst({ where: { Email: createUserDto.Email } });
 
     if (user) {
       return 'Tên người dùng đã được sử dụng';
@@ -45,7 +45,7 @@ export class UsersService {
   }
 
   async findOneUserName(username: string) {
-    const user = await this.prisma.users.findUnique({ where: { id: username } });
+    const user = await this.prisma.users.findFirst({ where: { Username: username } });
     if (!user) {
       console.log(`Không tìm thấy người dùng với tên người dùng ${username}`);
       return null;
@@ -55,7 +55,7 @@ export class UsersService {
   }
 
   async findOneUser(id: string) {
-    const user = await this.prisma.users.findUnique({ where: { id } });
+    const user = await this.prisma.users.findFirst({ where: { UserID: Number(id) } });
     if (!user) {
       console.log(`Không tìm thấy người dùng với ID = ${id}`);
       return null;
@@ -94,7 +94,7 @@ export class UsersService {
 
     if (updated) {
       await this.prisma.users.update({
-        where: { id },
+        where: { UserID: Number(id) },
         data: user,
       });
       console.log('Người dùng với ID: ', id, ' đã được cập nhật thành công');
@@ -112,7 +112,7 @@ export class UsersService {
       throw new NotFoundException(`Không tìm thấy người dùng với ID = ${id}`);
     }
 
-    await this.prisma.users.delete({ where: { id } });
+    await this.prisma.users.delete({ where: { UserID: Number(id) } });
     console.log(`Xóa người dùng có ID = ${id} thành công`);
     return `User with ID ${id} has been successfully deleted`;
   }
@@ -125,13 +125,12 @@ export class UsersService {
       return 'Người dùng không tồn tại';
     }
 
-    if (await bcrypt.compare(password, user.PasswordHash)) {
+    if (await bcrypt.compare(password, String(user.PasswordHash))) {
       console.log('Đăng nhập thành công');
 
-      // Tạo JWT token
       const payload = {
         username: user.Username,
-        sub: user.id,
+        sub: user.UserID,
         roles: user.Role,
       };
       const token = jwt.sign(payload, jwtConstants.secret, { expiresIn: '7d' });
@@ -165,7 +164,7 @@ export class UsersService {
 
     // Update user's password hash
     await this.prisma.users.update({
-      where: { id: user.id },
+      where: { UserID: user.UserID },
       data: { PasswordHash: hashedPassword },
     });
     console.log('Đổi mật khẩu thành công');
